requirements.txt
# 한글 선거 개표상황표 OCR 시스템 필수 패키지
# Korean Election Ballot OCR System - Required Packages

# Core OCR Engine
easyocr>=1.7.0

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0

# Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Excel Export
openpyxl>=3.1.0

# Optional: For better performance
# torch>=2.0.0  # GPU support (large package)
# torchvision>=0.15.0
